skl: some light refactoring
This is some light refactoring of the code to remove some duplicate code,
logically separate some code that was in the same unit files, and fixing
everything up from the refactoring.
