+++ event_log.c
@@ -17,6 +17,7 @@
  */
 
 #include <boot.h>
+#include <string.h>
 #include <tags.h>
 #include "tpmlib/tpm.h"
 #include "tpmlib/tpm2_constants.h"
