+++ include/types.h
@@ -38,8 +38,6 @@ typedef  int16_t s16;
 typedef  int32_t s32;
 typedef  int64_t s64;
 
-#include <string.h>	/* memcpy, memset */
-
 #else
 /*
  * If we are freestanding (i.e. building skl.bin itself), there is no
@@ -63,23 +61,6 @@ typedef long		ssize_t;
 typedef _Bool		bool;
 
 #define NULL ((void *)0)
-
-/*
- * Local declaration of bits of libc
- *
- * Use __builtin_???() wherever possible to allow the compiler to perform
- * optimisations (e.g. constant folding) where possible.  Calls to ???() will
- * be emitted as needed.
- */
-
-void *memset(void *s, int c, size_t n);
-#define memset(d, c, n) __builtin_memset(d, c, n)
-
-void *memcpy(void *dst, const void *src, size_t n);
-#define memcpy(d, s, n) __builtin_memcpy(d, s, n)
-
-size_t strlen(const char *s);
-#define strlen(s)       __builtin_strlen(s)
 
 #endif /* __STDC_HOSTED__ */
 #endif /* __TYPES_H__ */
