+++ tpmlib/tpm1_cmds.c
@@ -23,6 +23,8 @@
 
 #endif
 
+#include <string.h>
+
 #include "tpm.h"
 #include "tpmbuff.h"
 #include "tis.h"
