+++ tpmlib/tpm2_cmds.c
@@ -25,6 +25,8 @@
 
 #endif
 
+#include <string.h>
+
 #include "tpm.h"
 #include "tpmbuff.h"
 #include "tpm_common.h"
