+++ tpmlib/tpm_buff.c
@@ -19,6 +19,8 @@
 
 #endif
 
+#include <string.h>
+
 #include "tpm.h"
 #include "tpmbuff.h"
 #include "tpm_common.h"
